requirements.txt
Cython==0.29
ipython==7.1.1
Image==1.5.27
pyOpenSSL==18.0.0
Pillow==5.3.0
Babel==2.6.0
blinker==1.4
bpython==0.17.1
cryptography==2.4.1
cx_Oracle==7.0.0
python-dotenv==0.9.1
flask_babel==0.12.2
flask_babelex==0.9.3
protobuf==3.6.1
greenlet==0.4.15
ipaddr==2.2.0
libmc==1.1.0
lxml==4.2.5
python_memcached==1.59
mock==2.0.0
ndg_httpsclient==0.5.1
numpy==1.15.4
opcode==1534770263.106256
ordereddict==1.1
pretty==0.1
psycopg2==2.7.6.1
pyasn1==0.4.4
Pygments==2.2.0
pyinotify==0.9.6
pylibmc==1.6.0
redis==2.10.6
simplejson==3.16.0
sitecustomize==1534770371.8174736
sqlparse==0.2.4
tblib==1.3.2
termcolor==1.1.0
usercustomize==1.0.0
uwsgi==2.0.17.1
watchdog==0.9.0
wincertstore==0.2
wtforms_alchemy==0.16.7
PyYAML==3.13
